Avoid race conditions for resource monitoring.
It is possible that init finishes and successfully wait for the thread,
and only then is killed by the resource monitor. In pathological
cases, this might result in accidentally killing a new process which
got a recycled pid.

We also add a lock to make the resource monitor wait for the wait thread
to set up its signal handler. Otherwise, we might send a signal to the
thread that winds up killing the process.