Start support for contest-less pages.
This PR makes the following pages support contest-less:
- home
- login
- register
- problem view

Also, filters are updated to not be broken in this mode, in that one has
to explicitly set filters of various types. Without this fix, a filter
of e.g. the empty slice of problem IDs would list all problems rather
than no problems.